fix some demo playback/capture issues.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4887 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/protocol.h
@@ -1046,17 +1046,17 @@ typedef struct usercmd_s
 	qbyte	lightlevel;
 
 	//freestyle
-	int msec;
-	int buttons;
-	int weapon;
-	int servertime;
-	float fservertime;
-	float fclienttime;
+	float	msec;
+	int		buttons;
+	int		weapon;
+	int		servertime;
+	float	fservertime;
+	float	fclienttime;
 
-	vec2_t cursor_screen;
-	vec3_t cursor_start;
-	vec3_t cursor_impact;
-	int cursor_entitynumber;
+	vec2_t	cursor_screen;
+	vec3_t	cursor_start;
+	vec3_t	cursor_impact;
+	int		cursor_entitynumber;
 } usercmd_t;
 
 typedef struct q2usercmd_s
